imp: split pid_info code from imp/kill.c
Problem: The code that creates 'struct pid_info' is contained within
imp/kill.c, but would be better abstracted into its own file for
better organization, readabliity, and testability.

Move the code related to 'struct pid_info' into its own source file.
